modified Visualizer

+++ Global/util/util.py
@@ -62,9 +62,9 @@ class compute_performance():
 
     def update(self,preds,gts):
         preds = paddle.clip((preds + 1) * 127.5 + 0.5, 0, 255).transpose(
-            [0,2, 3, 1]).cast("uint8").numpy()
+            [0, 2, 3, 1]).cast("uint8").cast("float32").numpy()
         gts = paddle.clip((gts + 1) * 127.5 + 0.5, 0, 255).transpose(
-            [0, 2,3, 1]).cast("uint8").numpy()
+            [0, 2, 3, 1]).cast("uint8").cast("float32").numpy()
         self.PSNR.update(preds,gts)
         self.SSIM.update(preds,gts)
         self.FID.update(preds,gts)
